Test classes were written using Mockito and JUnit. Also, the backend logic service was improved, and the frontend design was updated.

+++ src/main/java/org/example/warehouseonline/entity/WareHouseItems.java
@@ -30,7 +30,7 @@ public class WareHouseItems {
 
     private byte[] image;
 
-    private Integer price;
+    private double price;
 
     private Integer quantity;
